Fixed small temporal issue
The system no longer outputs the last input as its new starting position. This was fixed as in the beginning of whatever input the user did, it would always output the last values in the same way. meaning. 'sand' and 'band' would output the same values for the character 'd'.