Changed reflection variable mutable path origin
Change the reflection variable to a u8 as it makes the reflection code much cleaner.

Made origin change to current position on each movement. this allows for the geometry to be fully explored more easily.
it also ensures that the beginning of the output sequence have much larger potential entropy.