Included the normalizer scirpt and rule system
Void Vault now support site specific rules. what this mean is that if a website does not support a set of characters or a length, you can configure the extension to modify the output of Void Vault to conform to these rules. It does this by removing characters not allowed, or collapses them into ascii. As the default, it allows for the whole UTF8 space.